Somewhat fix MSAN build (#3603)

+++ ci.sh
@@ -561,6 +561,7 @@ cmd_msanfuzz() {
   # Install msan if needed before changing the flags.
   detect_clang_version
   local msan_prefix="${HOME}/.msan/${CLANG_VERSION}"
+  # TODO(eustas): why libc++abi.a is bad?
   if [[ ! -d "${msan_prefix}" || -e "${msan_prefix}/lib/libc++abi.a" ]]; then
     # Install msan libraries for this version if needed or if an older version
     # with libc++abi was installed.
@@ -653,10 +654,16 @@ cmd_msan() {
   CMAKE_MODULE_LINKER_FLAGS+=" ${msan_linker_flags[@]}"
   CMAKE_SHARED_LINKER_FLAGS+=" ${msan_linker_flags[@]}"
   strip_dead_code
+
+  # MSAN share of stack size is non-negligible.
+  TEST_STACK_LIMIT="none"
+
+  # TODO(eustas): investigate why fuzzers do not link when MSAN libc++ is used
   cmake_configure "$@" \
     -DCMAKE_CROSSCOMPILING=1 -DRUN_HAVE_STD_REGEX=0 -DRUN_HAVE_POSIX_REGEX=0 \
     -DJPEGXL_ENABLE_TCMALLOC=OFF -DJPEGXL_WARNINGS_AS_ERRORS=OFF \
-    -DCMAKE_REQUIRED_LINK_OPTIONS="${msan_linker_flags[@]}"
+    -DCMAKE_REQUIRED_LINK_OPTIONS="${msan_linker_flags[@]}" \
+    -DJPEGXL_ENABLE_FUZZERS=OFF
   cmake_build_and_test
 }
 
@@ -665,6 +672,8 @@ cmd_msan() {
 cmd_msan_install() {
   local tmpdir=$(mktemp -d)
   CLEANUP_FILES+=("${tmpdir}")
+  local msan_root="${HOME}/.msan"
+  mkdir -p "${msan_root}"
   # Detect the llvm to install:
   export CC="${CC:-clang}"
   export CXX="${CXX:-clang++}"
@@ -672,23 +681,36 @@ cmd_msan_install() {
   # Allow overriding the LLVM checkout.
   local llvm_root="${LLVM_ROOT:-}"
   if [ -z "${llvm_root}" ]; then
-    local llvm_tag="llvmorg-${CLANG_VERSION}.0.0"
-    case "${CLANG_VERSION}" in
-      "6.0")
-        llvm_tag="llvmorg-6.0.1"
-        ;;
-      "7")
-        llvm_tag="llvmorg-7.0.1"
-        ;;
-    esac
-    local llvm_targz="${tmpdir}/${llvm_tag}.tar.gz"
-    curl -L --show-error -o "${llvm_targz}" \
-      "https://github.com/llvm/llvm-project/archive/${llvm_tag}.tar.gz"
+    declare -A llvm_tag_by_version=(
+      ["6.0"]="6.0.1"
+      ["7"]="7.1.0"
+      ["8"]="8.0.1"
+      ["9"]="9.0.2"
+      ["10"]="10.0.1"
+      ["11"]="11.1.0"
+      ["12"]="12.0.1"
+      ["13"]="13.0.1"
+      ["14"]="14.0.6"
+      ["15"]="15.0.7"
+      ["16"]="16.0.6"
+      ["17"]="17.0.6"
+      ["18"]="18.1.6"
+    ) 
+    local llvm_tag="${CLANG_VERSION}.0.0"
+    if [[ -n "${llvm_tag_by_version["${CLANG_VERSION}"]}" ]]; then
+      llvm_tag=${llvm_tag_by_version["${CLANG_VERSION}"]}
+    fi
+    llvm_tag="llvmorg-${llvm_tag}"
+    local llvm_targz="${msan_root}/${llvm_tag}.tar.gz"
+    if [ ! -f "${llvm_targz}" ]; then
+      curl -L --show-error -o "${llvm_targz}" \
+        "https://github.com/llvm/llvm-project/archive/${llvm_tag}.tar.gz"
+    fi
     tar -C "${tmpdir}" -zxf "${llvm_targz}"
     llvm_root="${tmpdir}/llvm-project-${llvm_tag}"
   fi
 
-  local msan_prefix="${HOME}/.msan/${CLANG_VERSION}"
+  local msan_prefix="${msan_root}/${CLANG_VERSION}"
   rm -rf "${msan_prefix}"
 
   local TARGET_OPTS=""
@@ -700,32 +722,29 @@ cmd_msan_install() {
     "
   fi
 
-  declare -A CMAKE_EXTRAS
-  CMAKE_EXTRAS[libcxx]="\
-    -DLIBCXX_CXX_ABI=libstdc++ \
-    -DLIBCXX_INSTALL_EXPERIMENTAL_LIBRARY=ON"
-
-  for project in libcxx; do
-    local proj_build="${tmpdir}/build-${project}"
-    local proj_dir="${llvm_root}/${project}"
-    mkdir -p "${proj_build}"
-    cmake -B"${proj_build}" -H"${proj_dir}" \
-      -G Ninja \
-      -DCMAKE_BUILD_TYPE=Release \
-      -DLLVM_USE_SANITIZER=Memory \
-      -DLLVM_PATH="${llvm_root}/llvm" \
-      -DLLVM_CONFIG_PATH="$(which llvm-config llvm-config-7 llvm-config-6.0 | \
-                            head -n1)" \
-      -DCMAKE_CXX_FLAGS="${CMAKE_CXX_FLAGS}" \
-      -DCMAKE_C_FLAGS="${CMAKE_C_FLAGS}" \
-      -DCMAKE_EXE_LINKER_FLAGS="${CMAKE_EXE_LINKER_FLAGS}" \
-      -DCMAKE_SHARED_LINKER_FLAGS="${CMAKE_SHARED_LINKER_FLAGS}" \
-      -DCMAKE_INSTALL_PREFIX="${msan_prefix}" \
-      ${TARGET_OPTS} \
-      ${CMAKE_EXTRAS[${project}]}
-    cmake --build "${proj_build}"
-    ninja -C "${proj_build}" install
-  done
+  local build_dir="${tmpdir}/build-llvm"
+  mkdir -p "${build_dir}"
+  cd ${llvm_root}
+  cmake -B"${build_dir}" \
+    -G Ninja \
+    -S runtimes \
+    -DCMAKE_BUILD_TYPE=Release \
+    -DLLVM_USE_SANITIZER=Memory \
+    -DLLVM_ENABLE_RUNTIMES="libcxx;libcxxabi;libunwind;compiler-rt" \
+    -DLIBCXXABI_ENABLE_SHARED=ON \
+    -DLIBCXXABI_ENABLE_STATIC=OFF \
+    -DLIBCXX_ENABLE_SHARED=ON \
+    -DLIBCXX_ENABLE_STATIC=OFF \
+    -DCMAKE_CXX_FLAGS="${CMAKE_CXX_FLAGS}" \
+    -DCMAKE_C_FLAGS="${CMAKE_C_FLAGS}" \
+    -DCMAKE_EXE_LINKER_FLAGS="${CMAKE_EXE_LINKER_FLAGS}" \
+    -DCMAKE_SHARED_LINKER_FLAGS="${CMAKE_SHARED_LINKER_FLAGS}" \
+    -DCMAKE_INSTALL_PREFIX="${msan_prefix}" \
+    -DLLVM_PATH="${llvm_root}/llvm" \
+    -DLLVM_CONFIG_PATH="$(which llvm-config-${CLANG_VERSION} llvm-config | head -n1)" \
+     ${TARGET_OPTS}
+  cmake --build "${build_dir}"
+  ninja -C "${build_dir}" install
 }
 
 # Internal build step shared between all cmd_ossfuzz_* commands.
